Updated README, added JSON file, and Re-Wrote logic in app.js
