+++ README.md
@@ -1,2 +1,74 @@
-# kindIdeas
-More details coming soon.
+<center>
+    <img src="./app/assets/icon.png" alt="Kind Ideas">
+</center>
+
+<center>Spreading kind ideas One JSON String at a time!</center>
+
+<br/>
+
+<center>
+
+[![Build Status](https://travis-ci.com/username/projectname.svg?branch=master)](https://travis-ci.com/username/projectname) &nbsp; &nbsp;
+[![License: MIT](https://img.shields.io/badge/License-MIT-yellow.svg)](https://opensource.org/licenses/MIT)
+
+</center>
+<br> <br>
+
+# Overview – 
+KindIdeas is an open source project that displays Kind things to do.
+
+<br/>
+
+# How to Contribute –
+Contributions to kindIdeas is welcomed! The vision of this project, was to have several people contribute kind ideas that'll make the world a kinder place! 
+
+Our <b>Contribution Guidelines</b> are avaibale to <b>read bellow...</b> We ask you give them a peek, before making a contribution!
+
+<br/>
+
+## Contribution Guidelines –
+
+### <b>Bug Reporting:</b>
+This section will guide you through on how to open a bug report for Kind Ideas. Please follow our guidelines, as they help us understand the issue quickly! 
+
+<b>Before you submit a Bug Report:</b>
+Before you submit a bug report, please consider the following...<br/>
+1. Has the bug already been descovered? Check the Issues tab of the project, so duplicate bug reports are not created.
+
+2. If the bug occured in browser, please be sure you're using the most updated version – if you are – try using a different one to duplicate the bug.
+
+3. Collect as much details as you can. Screen grabs, videos, and a written report are great to have while debugging an issue!
+
+<br>
+<b>How to Submit a Bug Report:</b>
+If you were able to duplicate the bug, and believe a bug report is warrented the following will walk you through creating a bug report. Bug are tracked using GitHub issues. Create a new issue and follow the template provided. <br/> <br/>
+
+
+### <b>Suggesting Improvements:</b>
+Improvement suggestions are tracked using <b>GitHub Issues</b> to submit a feature request, or minor imporvements create a new issue following the steps bellow... <br/>
+
+1. <b>Suggesting a New Feature –</b> To suggest a new feature, open an issue on GitHub Issues. Provide details of the new feature, you'd like to see implemented. Label it a "feature request". And then submit the issue. We'll take a look, and let you know if we decide to move forword!
+
+2. <b>Suggesting a Minior Improvement – </b> To suggest a minor improvement, open a new issue on GitHub Issues. Provide the details of the improvement you'd like to see implmented, and label it "minor suggestion". We'll take a look, and let you know if we decide to move forword! <br/> <br/>
+
+### <b>First Time Contribution:</b>
+Not sure where to begin? Start looking through "good first issues" in the GitHub Issues tab.
+
+<b>Local Project Dependencies – </b> Kind Ideas uses several local packages that may need to be installed on your machine, before you can begin developing! They're listed bellow...
+
+1. Node.js 
+2. Node Package Manager (NPM)
+3. http-server | More Details can be found here: https://github.com/indexzero/http-server <br/> <br/>
+
+
+
+<br/>
+
+# The Full Story –
+Kind Ideas was a simple thought that posed to my mind after watching the Ellen Show. She ends each of her shows by saying: “Be Kind to One Another.” 
+After hearing her say this phrase for the 100th time, it finally clicked for me. We all can say how we're doing great things, we can say that we give to charity,
+and we can claim that we’re making a difference. But is any of that doing any good? (Hear me out!) When we give to charity are we being kind to those people
+who we are giving to? If you give to a homeless shelter, yet yell and make fun of the homeless on the streets—what good did your giving really even do? 
+
+For many of us, we’re busy. We’re too busy to come up with a way to be Kind to another Human being! 
+This is why I decided to use technology to solve a problem that I strongly believe in. I built a website, that is essentially a Kind Ideas Generator. If someone needs help coming up with something nice to say to someone, or needs an idea of what to give someone — it is my hope this website will help!
