+++ app/assets/app.test.js
@@ -1,0 +1,2 @@
+const app = require('./app')
+
