+++ app/assets/style.css
@@ -16,6 +16,7 @@ h1{
 footer{
     margin-top: 150px;
 }
+
 button{
     width: 250px;
     height: 50px;
