+++ app/index.html
@@ -5,8 +5,8 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <title>Kind Ideas</title>
     <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link href="https://fonts.googleapis.com/css?family=Montserrat" rel="stylesheet">     <link rel="stylesheet" type="text/css" media="screen" href="style.css" />
-    <script src="main.js"></script>
+    <link href="https://fonts.googleapis.com/css?family=Montserrat" rel="stylesheet">     <link rel="stylesheet" type="text/css" media="screen" href="./assets/style.css" />
+    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script>
 </head>
 <body>
     <h1>Kindness Generator</h1>
@@ -20,11 +20,11 @@
     <button onclick="newDeed(); hidePlaceholder();">Get a New Kind Act</button>
 
     <footer>
-        <p>This project was created by <a href="http://www.adammcknight.co" target="_blank">Adam McKnight</a>
+        <p>This project was created by <a href="https://mcknight.herokuapp.com" target="_blank">Adam McKnight</a>
         &mdash; &copy; 2018 All Rights Reserved.
         </p>
     </footer>
 
-    <script src="application.js"></script>
+    <script src="./assets/app.js"></script>
 </body>
 </html>
